CliParser Class Refactor Changed from parser::Parser to cliparser::CliParser Fixed missing /include on /source/cliParser/CMakeLists
