+++ src/engine/test/CMakeLists.txt
@@ -26,4 +26,4 @@ add_subdirectory(${TEST_SOURCE_DIR}/json)
 add_subdirectory(${TEST_SOURCE_DIR}/builder)
 add_subdirectory(${TEST_SOURCE_DIR}/graph)
 add_subdirectory(${TEST_SOURCE_DIR}/catalog)
-add_subdirectory(${TEST_SOURCE_DIR}/argumentParser)
+add_subdirectory(${TEST_SOURCE_DIR}/cliParser)
